+++ CHANGELOG.md
@@ -27,7 +27,8 @@
 - Added uuid API validation
 - Implemented OCLC Classify Process v2
 - Added integration tests for UofM ingestion process
-- Updated fulfill process and script to check rights status before updating manifest
+- Updated fulfill process to check rights status before updating manifest
+- Deleted fulfill script due to no longer being necessary
 
 ## Fixed
 - Changed HATHI_DATAFILES outdated link in development, example, and local yaml files